+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Reset Turnstile captcha on auth error
+
 ## [0.9.1] - 2025-02-27
 
 ### Changed
@@ -14,7 +17,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Added Turnstile support to ChangePassword component to fix password change failures when Turnstile is enabled
-- Reset Turnstile captcha on auth error
 
 ## [0.9.0] - 2025-02-27